requirements.txt
fastapi~=0.121.0
pydantic~=2.12.4
uvicorn
bcrypt~=5.0.0
python-jose~=3.5.0
python-multipart
mariadb~=1.1.14
jose~=1.0.0
starlette~=0.49.3
python-dateutil

